pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    { name = "TdewT", email = "szwej.jakub@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "colorlog (>=6.10.1,<7.0.0)",
    "cassandra-repositories @ git+https://github.com/ZPI-KKM-WSIZ/database.git@main#subdirectory=repositories",
    "pytest (>=9.0.2,<10.0.0)",
    "httpx (>=0.28.1,<0.29.0)"
]

[tool.poetry]
package-mode = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)"
]
